updated claims with jwt logic

+++ controller/claims.js
@@ -1,12 +1,22 @@
 const express = require("express");
 const mongoose = require("mongoose");
 const ProjectExpenseClaims = require("../models/projectExpenseClaimModel");
+const { access } = require("fs");
+const auth = require("../auth/auth");
 
 // Delete project expense claims
 
 const deleteProjectExpenseClaims = async (req, res) => {
   try {
     const filter = req.query.claimId;
+    const appears = await ProjectExpenseClaims.findOne({ claimId: filter });
+    const accessToken = req.body.accessToken;
+    if (!appears) {
+      return res.status(404).json({ message: "There is no such claim Id." });
+    }
+    if (!auth.verify_credentials(accessToken, appears.employeeId)) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
     const projectExpenseClaims = await ProjectExpenseClaims.findOneAndDelete({
       claimId: filter,
     });
@@ -30,6 +40,11 @@ const deleteProjectExpenseClaims = async (req, res) => {
 const updateProjectExpenseClaims = async (req, res) => {
   try {
     const filter = req.body.claimId;
+    const employeeId = req.body.employeeId;
+    const accessToken = req.body.accessToken;
+    if (!auth.verify_credentials(accessToken, employeeId)) {
+      return res.status(401).json({ message: "Unauthorized" });
+    }
     const expense = await ProjectExpenseClaims.findOneAndUpdate(
       { claimId: filter },
       req.body,
